refactor planet v2 analysis test
