+++ ignite/pkg/cosmosanalysis/module/testdata/planet_v2/proto/planet/mars/mars.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
-package tendermint.planet.planet;
+package tendermint.planet.mars;
 import "google/api/annotations.proto";
-option go_package = "github.com/tendermint/mars/x/mars/types";
+option go_package = "github.com/tendermint/planet/x/mars/types";
 
 service Query {
   rpc MyQuery(QueryMyQueryRequest) returns (QueryMyQueryResponse) {
