+++ ignite/pkg/cosmosanalysis/module/testdata/planet_v2/x/mars/keeper/grpc_query_my_query.go
@@ -3,7 +3,7 @@ package keeper
 import (
 	"context"
 
-	"github.com/tendermint/planet/x/planet/types"
+	"github.com/tendermint/planet/x/mars/types"
 )
 
 type Keeper struct{}
